requirements.txt
Django==4.2.17
django-cors-headers==3.13.0
django-crispy-forms==1.14.0
django-filter==22.1
django-prometheus==2.2.0
django-redis==5.2.0
django-rest-swagger==2.2.0
djangorestframework==3.14.0
djangorestframework-simplejwt==5.2.2
drf-standardized-errors==0.14.1
django_mysql==4.9.0
gevent==24.2.1
mysqlclient==2.1.1
# mysql-connector-python==8.0.32
numpy==1.24.0
redis==4.5.5
python-logstash==0.4.8
minio==7.1.12
requests==2.32.0
shortuuid==1.0.11
python-telegram-bot==13.12
pathvalidate==2.5.2
gunicorn==22.0.0
pydantic==2.7.4
django-mysql==4.9.0
strenum==0.4.8
pymongo==4.6.3
python-dateutil==2.8.2
mongoengine==0.27.0
ffmpeg-python==0.2.0
pandas==2.2.1
openpyxl==3.1.2
ecs-logging==2.1.0
kafka-python==2.0.2
pytelegraf==0.3.3
Pillow==10.3.0

#langchain-core==0.3.23
#langchain==0.3.10
#langchain-community==0.3.10
#langchain-openai==0.2.12
#langchain-ollama==0.2.1
#langchain-text-splitters==0.3.2
#langgraph==0.2.56

langchain==0.3.1
langchain-community==0.3.1
langchain-core==0.3.6
langchain-openai==0.2.1
langchain-text-splitters==0.3.0
